PIM-5952: Add default message in empty family filter
