+++ CHANGELOG-1.6.md
@@ -10,6 +10,7 @@
 - PIM-5959: Change wording of XLSX to Excel on mass actions
 - PIM-5938: Fix typos in import/export tooltips
 - PIM-5966: Fix style about category filter on product export builder
+- PIM-5952: Add the information "No condition on families" in the family field of the Export Builder
 
 # 1.6.0 (2016-08-30)
