+++ features/Context/Page/Base/Base.php
@@ -32,6 +32,10 @@ class Base extends Page
         'Navigation Bar'   => ['css' => 'header#oroplatform-header'],
         'Container'        => ['css' => '#container'],
         'Locales dropdown' => ['css' => '#locale-switcher'],
+        'Tabs'             => ['css' => '#form-navbar'],
+        'Oro tabs'         => ['css' => '.navbar.scrollspy-nav'],
+        'Form tabs'        => ['css' => '.nav-tabs.form-tabs'],
+        'Active tab'       => ['css' => '.form-horizontal .tab-pane.active'],
     ];
 
     /**
@@ -396,5 +400,101 @@ class Base extends Page
         $session->buttondown('');
         $session->moveto(['element' => $to->getID()]);
         $session->buttonup('');
+    }
+
+    /**
+     * Visit the specified tab
+     *
+     * @param string $tab
+     */
+    public function visitTab($tab)
+    {
+        $tabs = $this->getPageTabs();
+
+        $tabDom = $this->spin(function () use ($tabs, $tab) {
+            return $tabs->findLink($tab);
+        }, sprintf('Could not find a tab named "%s"', $tab));
+
+        $this->spin(function () {
+            $loading = $this->find('css', '#loading-wrapper');
+
+            return null === $loading || !$loading->isVisible();
+        }, sprintf('Could not visit tab %s because of loading wrapper', $tab));
+
+        $this->spin(function () use ($tabDom) {
+            $tabDom->click();
+
+            return $tabDom->getParent()->hasClass('active') || $tabDom->getParent()->hasClass('tab-scrollable');
+        }, sprintf('Cannot switch to the tab %s', $tab));
+    }
+
+    /**
+     * Get the tabs in the current page
+     *
+     * @return NodeElement[]
+     */
+    public function getTabs()
+    {
+        $tabs = $this->spin(function () {
+            return $this->find('css', $this->elements['Tabs']['css']);
+        }, sprintf('Cannot find "%s" element', $this->elements['Tabs']['css']));
+
+        if (null === $tabs) {
+            $tabs = $this->getElement('Oro tabs');
+        }
+
+        return $tabs->findAll('css', 'a');
+    }
+
+    /**
+     * Get the form tab containg $tab text
+     *
+     * @param string $tab
+     *
+     * @return NodeElement|null
+     */
+    public function getFormTab($tab)
+    {
+        $tabs = $this->getPageTabs();
+
+        try {
+            $node = $this->spin(function () use ($tabs, $tab) {
+                return $tabs->find('css', sprintf('a:contains("%s")', $tab));
+            }, sprintf('Cannot find form tab "%s"', $tab));
+        } catch (\Exception $e) {
+            $node = null;
+        }
+
+        return $node;
+    }
+
+    /**
+     * Get the specified tab
+     *
+     * @return NodeElement
+     */
+    public function getTab($tab)
+    {
+        return $this->find('css', sprintf('a:contains("%s")', $tab));
+    }
+
+    /**
+     * Returns the tabs of the current page, if any.
+     *
+     * @return NodeElement
+     */
+    protected function getPageTabs()
+    {
+        return $this->spin(function () {
+            $tabs = $this->find('css', $this->elements['Tabs']['css']);
+            if (null === $tabs) {
+                $tabs = $this->find('css', $this->elements['Oro tabs']['css']);
+            }
+            if (null === $tabs) {
+                $tabs = $this->find('css', $this->elements['Form tabs']['css']);
+            }
+
+            return $tabs;
+        }, 'Cannot find any tabs in this page');
     }
 }
